requirements.txt
Babel==2.12.1
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
chardet==4.0.0
charset-normalizer==3.4.2
click==8.1.8
dnspython==2.7.0
docutils==0.18.1
dotenv==0.9.9
Flask==3.1.1
flask-cors==6.0.1
idna==3.3
imagesize==1.3.0
importlib_metadata==8.7.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
olefile==0.46
packaging==21.3
Pillow==11.0.0
platformdirs==4.2.0
pluggy==1.0.0
Pygments==2.14.0
pymongo==4.13.2
pyparsing==3.0.6
pytest==7.3.0
python-dotenv==1.1.1
railroad-diagrams==3.0.1
requests==2.31.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.7
Sphinx==5.3.0
sphinxcontrib-serializinghtml==1.1.5
toml==0.10.2
typing_extensions==4.14.0
urllib3==1.26.7
Werkzeug==3.1.3
zipp==3.23.0

# --- Added for recommender system ---
numpy==1.23.5
pandas==2.2.3
scikit-surprise==1.1.3
scikit-learn==1.5.2
